requirements.txt
# Authentication Microservice - Enterprise Security Grade
# Python 3.12+ Required

# FastAPI & ASGI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database (PostgreSQL 16 + Async)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# Security
passlib[argon2]>=1.7.4
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0
pyotp>=2.9.0

# UUID v7 Support
uuid6>=2023.5.2

# AWS Integration
boto3>=1.29.0

# Redis (Rate Limiting)
redis>=5.0.0
hiredis>=2.2.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Email Validation
email-validator>=2.1.0

# QR Code Generation
qrcode[pil]>=7.4.2

# Optional: Development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0
